added rows and create assignment

+++ src/Logic/TaskLogic.ts
@@ -4,18 +4,26 @@ import ICanvas from 'src/canvas/ICanvas';
 import DB from 'src/Data/db';
 import ITaskLogic from './ITaskLogic';
 
-export class TaskLogic implements ITaskLogic {
+export default class TaskLogic implements ITaskLogic {
   constructor(
     private readonly canvasService: ICanvas,
     private readonly dataBaseService: DB,
   ) {}
 
   async GetTasks(id: string) {
-    return await this.dataBaseService.GetTasks(id);
+    let tasks = await this.dataBaseService.GetTasks(id);
+    if(tasks.length == 0){
+      console.log("No tasks found")
+      return false;
+    }
+    return tasks;
   }
 
   async CreateTask(task: CreateTaskBody) {
-    console.log(task);
+    if(task.AssignmentId == undefined || task.Status == undefined || task.Name == undefined || task.DueDate == undefined){
+      return false;
+    }
+    
     return await this.dataBaseService.CreateTasks(
       randomUUID(),
       task.AssignmentId,
@@ -26,6 +34,14 @@ export class TaskLogic implements ITaskLogic {
   }
 
   async Edit(id: string, task: CreateTaskBody) {
+    if(task.AssignmentId == undefined || task.Status == undefined || task.Name == undefined || task.DueDate == undefined){
+      return false;
+    }
+
+    if(await this.dataBaseService.DoesTaskExist(id) < 1){
+      return false;
+    }
+
     return await this.dataBaseService.EditTask(
       id,
       task.AssignmentId,
@@ -36,6 +52,10 @@ export class TaskLogic implements ITaskLogic {
   }
 
   async Delete(id: string) {
+    if(await this.dataBaseService.DoesTaskExist(id) < 1){
+      return false;
+    }
+    
     return await this.dataBaseService.DeleteTask(id);
   }
 }
